Apply a function to each element of a vector via futures
Source: R/future_map.R

These functions work exactly the same as purrr::map() functions, but allow you to run the map in parallel. There are a number of future.* arguments to allow you to fine tune the parallel processing. The documentation is adapted from both purrr::map(), and future.apply::future_lapply(), so look there for more details.

future_map(.x, .f, ..., future.globals = TRUE, future.packages = NULL,
  future.seed = FALSE, future.lazy = FALSE, future.scheduling = 1)

future_map_chr(.x, .f, ..., future.globals = TRUE, future.packages = NULL,
  future.seed = FALSE, future.lazy = FALSE, future.scheduling = 1)

future_map_dbl(.x, .f, ..., future.globals = TRUE, future.packages = NULL,
  future.seed = FALSE, future.lazy = FALSE, future.scheduling = 1)

future_map_int(.x, .f, ..., future.globals = TRUE, future.packages = NULL,
  future.seed = FALSE, future.lazy = FALSE, future.scheduling = 1)

future_map_lgl(.x, .f, ..., future.globals = TRUE, future.packages = NULL,
  future.seed = FALSE, future.lazy = FALSE, future.scheduling = 1)

future_map_dfr(.x, .f, ..., .id = NULL, future.globals = TRUE,
  future.packages = NULL, future.seed = FALSE, future.lazy = FALSE,
  future.scheduling = 1)

future_map_dfc(.x, .f, ..., future.globals = TRUE, future.packages = NULL,
  future.seed = FALSE, future.lazy = FALSE, future.scheduling = 1)

future_map_if(.x, .p, .f, ..., future.globals = TRUE,
  future.packages = NULL, future.seed = FALSE, future.lazy = FALSE,
  future.scheduling = 1)

future_map_at(.x, .at, .f, ..., future.globals = TRUE,
  future.packages = NULL, future.seed = FALSE, future.lazy = FALSE,
  future.scheduling = 1)

Arguments

.x
A list or atomic vector.

.f
A function, formula, or atomic vector.

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. Within a list, wrap strings in get-attr() to extract named attributes. If a component is not present, the value of .default will be returned.

...
Additional arguments passed on to .f.

future.globals
A logical, a character vector, or a named list for controlling how globals are handled. For details, see below section.

future.packages
(optional) a character vector specifying packages to be attached in the R environment evaluating the future.

future.seed
A logical or an integer (of length one or seven), or a list of length(.x) with pre-generated random seeds. For details, see below section.

future.lazy
Specifies whether the futures should be resolved lazily or eagerly (default).

future.scheduling
Average number of futures ("chunks") per worker. If 0.0, then a single future is used to process all elements of .x. If 1.0 or TRUE, then one future per worker is used. If 2.0, then each worker will process two futures (if there are enough elements in .x). If Inf or FALSE, then one future per element of .x is used.

.id
If not NULL a variable with this name will be created giving either the name or the index of the data frame.

.p
A single predicate function, a formula describing such a predicate function, or a logical vector of the same length as .x. Alternatively, if the elements of .x are themselves lists of objects, a string indicating the name of a logical element in the inner lists. Only those elements where .p evaluates to TRUE will be modified.

.at
A character vector of names or a numeric vector of positions. Only those elements corresponding to .at will be modified.

Value

All functions return a vector the same length as .x.

future_map() returns a list, future_map_lgl() a logical vector, future_map_int() an integer vector, future_map_dbl() a double vector, and future_map_chr() a character vector. The output of .f will be automatically typed upwards, e.g. logical -> integer -> double -> character.

Global variables

Argument future.globals may be used to control how globals should be handled similarly how the globals argument is used with future(). Since all function calls use the same set of globals, this function can do any gathering of globals upfront (once), which is more efficient than if it would be done for each future independently. If TRUE, NULL or not is specified (default), then globals are automatically identified and gathered. If a character vector of names is specified, then those globals are gathered. If a named list, then those globals are used as is. In all cases, .f and any ... arguments are automatically passed as globals to each future created as they are always needed.

Reproducible random number generation (RNG)

Unless future.seed = FALSE, this function guarantees to generate the exact same sequence of random numbers given the same initial seed / RNG state - this regardless of type of futures and scheduling ("chunking") strategy.

RNG reproducibility is achieved by pregenerating the random seeds for all iterations (over .x) by using L'Ecuyer-CMRG RNG streams. In each iteration, these seeds are set before calling .f(.x[[ii]], ...). Note, for large length(.x) this may introduce a large overhead. As input (future.seed), a fixed seed (integer) may be given, either as a full L'Ecuyer-CMRG RNG seed (vector of 1+6 integers) or as a seed generating such a full L'Ecuyer-CMRG seed. If future.seed = TRUE, then .Random.seed is returned if it holds a L'Ecuyer-CMRG RNG seed, otherwise one is created randomly. If future.seed = NA, a L'Ecuyer-CMRG RNG seed is randomly created. If none of the function calls .f(.x[[ii]], ...) uses random number generation, then future.seed = FALSE may be used.

In addition to the above, it is possible to specify a pre-generated sequence of RNG seeds as a list such that length(future.seed) == length(.x) and where each element is an integer seed that can be assigned to .Random.seed. Use this alternative with caution. Note that as.list(seq_along(.x)) is not a valid set of such .Random.seed values.

In all cases but future.seed = FALSE, the RNG state of the calling R processes after this function returns is guaranteed to be "forwarded one step" from the RNG state that was before the call and in the same way regardless of future.seed, future.scheduling and future strategy used. This is done in order to guarantee that an R script calling future_map() multiple times should be numerically reproducible given the same initial seed.

Examples

library(furrr)
library(dplyr) # for the pipe

plan(multiprocess)

1:10 %>%
  future_map(rnorm, n = 10) %>%
  future_map_dbl(mean)

#>  [1] 1.171826 1.860167 2.928960 3.858197 4.731077 5.852965 7.152533 8.817512
#>  [9] 8.942318 9.566968

# If each element of the output is a data frame, use
# map_dfr to row-bind them together:
mtcars %>%
  split(.$cyl) %>%
  future_map(~ lm(mpg ~ wt, data = .x)) %>%
  future_map_dfr(~ as.data.frame(t(as.matrix(coef(.)))))

#>   (Intercept)        wt
#> 1    39.57120 -5.647025
#> 2    28.40884 -2.780106
#> 3    23.86803 -2.192438