Map over multiple inputs simulatneously via futures
Source: R/future_map2.R

These functions work exactly the same as purrr::map2() functions, but allow you to run the map in parallel. Note that "parallel" as described in purrr is just saying that you are working with multiple inputs, and parallel in this case means that you can work on multiple inputs AND process them all in parallel as well.

future_map2(.x, .y, .f, ..., future.globals = TRUE, future.packages = NULL,
  future.seed = FALSE, future.lazy = FALSE, future.scheduling = 1)

future_map2_chr(.x, .y, .f, ..., future.globals = TRUE,
  future.packages = NULL, future.seed = FALSE, future.lazy = FALSE,
  future.scheduling = 1)

future_map2_dbl(.x, .y, .f, ..., future.globals = TRUE,
  future.packages = NULL, future.seed = FALSE, future.lazy = FALSE,
  future.scheduling = 1)

future_map2_int(.x, .y, .f, ..., future.globals = TRUE,
  future.packages = NULL, future.seed = FALSE, future.lazy = FALSE,
  future.scheduling = 1)

future_map2_lgl(.x, .y, .f, ..., future.globals = TRUE,
  future.packages = NULL, future.seed = FALSE, future.lazy = FALSE,
  future.scheduling = 1)

future_map2_dfr(.x, .y, .f, ..., .id = NULL, future.globals = TRUE,
  future.packages = NULL, future.seed = FALSE, future.lazy = FALSE,
  future.scheduling = 1)

future_map2_dfc(.x, .y, .f, ..., future.globals = TRUE,
  future.packages = NULL, future.seed = FALSE, future.lazy = FALSE,
  future.scheduling = 1)

Arguments

.x
Vectors of the same length. A vector of length 1 will be recycled.

.y
Vectors of the same length. A vector of length 1 will be recycled.

.f
A function, formula, or atomic vector.

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. Within a list, wrap strings in get-attr() to extract named attributes. If a component is not present, the value of .default will be returned.

...
Additional arguments passed on to .f.

future.globals
A logical, a character vector, or a named list for controlling how globals are handled. For details, see below section.

future.packages
(optional) a character vector specifying packages to be attached in the R environment evaluating the future.

future.seed
A logical or an integer (of length one or seven), or a list of length(.x) with pre-generated random seeds. For details, see below section.

future.lazy
Specifies whether the futures should be resolved lazily or eagerly (default).

future.scheduling
Average number of futures ("chunks") per worker. If 0.0, then a single future is used to process all elements of .x. If 1.0 or TRUE, then one future per worker is used. If 2.0, then each worker will process two futures (if there are enough elements in .x). If Inf or FALSE, then one future per element of .x is used.

.id
If not NULL a variable with this name will be created giving either the name or the index of the data frame.

Value

An atomic vector, list, or data frame, depending on the suffix. Atomic vectors and lists will be named if .x or the first element of .l is named.

If all input is length 0, the output will be length 0. If any input is length 1, it will be recycled to the length of the longest.

Global variables

Argument future.globals may be used to control how globals should be handled similarly how the globals argument is used with future(). Since all function calls use the same set of globals, this function can do any gathering of globals upfront (once), which is more efficient than if it would be done for each future independently. If TRUE, NULL or not is specified (default), then globals are automatically identified and gathered. If a character vector of names is specified, then those globals are gathered. If a named list, then those globals are used as is. In all cases, .f and any ... arguments are automatically passed as globals to each future created as they are always needed.

Reproducible random number generation (RNG)

Unless future.seed = FALSE, this function guarantees to generate the exact same sequence of random numbers given the same initial seed / RNG state - this regardless of type of futures and scheduling ("chunking") strategy.

RNG reproducibility is achieved by pregenerating the random seeds for all iterations (over .x) by using L'Ecuyer-CMRG RNG streams. In each iteration, these seeds are set before calling .f(.x[[ii]], ...). Note, for large length(.x) this may introduce a large overhead. As input (future.seed), a fixed seed (integer) may be given, either as a full L'Ecuyer-CMRG RNG seed (vector of 1+6 integers) or as a seed generating such a full L'Ecuyer-CMRG seed. If future.seed = TRUE, then .Random.seed is returned if it holds a L'Ecuyer-CMRG RNG seed, otherwise one is created randomly. If future.seed = NA, a L'Ecuyer-CMRG RNG seed is randomly created. If none of the function calls .f(.x[[ii]], ...) uses random number generation, then future.seed = FALSE may be used.

In addition to the above, it is possible to specify a pre-generated sequence of RNG seeds as a list such that length(future.seed) == length(.x) and where each element is an integer seed that can be assigned to .Random.seed. Use this alternative with caution. Note that as.list(seq_along(.x)) is not a valid set of such .Random.seed values.

In all cases but future.seed = FALSE, the RNG state of the calling R processes after this function returns is guaranteed to be "forwarded one step" from the RNG state that was before the call and in the same way regardless of future.seed, future.scheduling and future strategy used. This is done in order to guarantee that an R script calling future_map() multiple times should be numerically reproducible given the same initial seed.

Examples

library(furrr)
plan(multiprocess)

x <- list(1, 10, 100)
y <- list(1, 2, 3)

future_map2(x, y, ~ .x + .y)

#> [[1]]
#> [1] 2
#> 
#> [[2]]
#> [1] 12
#> 
#> [[3]]
#> [1] 103
#> 

# Split into pieces, fit model to each piece, then predict
by_cyl <- split(mtcars, mtcars$cyl)
mods <- future_map(by_cyl, ~ lm(mpg ~ wt, data = .))
future_map2(mods, by_cyl, predict)

#> $`4`
#>     Datsun 710      Merc 240D       Merc 230       Fiat 128    Honda Civic 
#>       26.47010       21.55719       21.78307       27.14774       30.45125 
#> Toyota Corolla  Toyota Corona      Fiat X1-9  Porsche 914-2   Lotus Europa 
#>       29.20890       25.65128       28.64420       27.48656       31.02725 
#>     Volvo 142E 
#>       23.87247 
#> 
#> $`6`
#>      Mazda RX4  Mazda RX4 Wag Hornet 4 Drive        Valiant       Merc 280 
#>       21.12497       20.41604       19.47080       18.78968       18.84528 
#>      Merc 280C   Ferrari Dino 
#>       18.84528       20.70795 
#> 
#> $`8`
#>   Hornet Sportabout          Duster 360          Merc 450SE          Merc 450SL 
#>            16.32604            16.04103            14.94481            15.69024 
#>         Merc 450SLC  Cadillac Fleetwood Lincoln Continental   Chrysler Imperial 
#>            15.58061            12.35773            11.97625            12.14945 
#>    Dodge Challenger         AMC Javelin          Camaro Z28    Pontiac Firebird 
#>            16.15065            16.33700            15.44907            15.43811 
#>      Ford Pantera L       Maserati Bora 
#>            16.91800            16.04103 
#>