Apply a function to each element of a vector, and its index via futures

These functions work exactly the same as purrr::imap() functions, but allow you to map in parallel.

future_imap(.x, .f, ..., .progress = FALSE, .options = future_options())

future_imap_chr(.x, .f, ..., .progress = FALSE, .options = future_options())

future_imap_dbl(.x, .f, ..., .progress = FALSE, .options = future_options())

future_imap_int(.x, .f, ..., .progress = FALSE, .options = future_options())

future_imap_lgl(.x, .f, ..., .progress = FALSE, .options = future_options())

future_imap_dfr(.x, .f, ..., .id = NULL, .progress = FALSE,
  .options = future_options())

future_imap_dfc(.x, .f, ..., .progress = FALSE, .options = future_options())

future_iwalk(.x, .f, ..., .progress = FALSE, .options = future_options())

Arguments

.x
A list or atomic vector.

.f
A function, formula, or atomic vector.

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. Within a list, wrap strings in get-attr() to extract named attributes. If a component is not present, the value of .default will be returned.

...
Additional arguments passed on to .f.

.progress
A logical, for whether or not to print a progress bar for multiprocess, multisession, and multicore plans.

.options
The future specific options to use with the workers. This must be the result from a call to future_options().

.id
If not NULL a variable with this name will be created giving either the name or the index of the data frame.

Value

A vector the same length as .x.

Examples

library(furrr)

plan(multiprocess)

future_imap_chr(sample(10), ~ paste0(.y, ": ", .x))

#>  [1] "1: 1"  "2: 8"  "3: 5"  "4: 2"  "5: 10" "6: 3"  "7: 7"  "8: 6"  "9: 4" 
#> [10] "10: 9"