Invoke functions via futures

These functions work exactly the same as purrr::invoke_map() functions, but allow you to invoke in parallel.

future_invoke_map(
  .f,
  .x = list(NULL),
  ...,
  .env = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = deprecated()
)

future_invoke_map_chr(
  .f,
  .x = list(NULL),
  ...,
  .env = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = deprecated()
)

future_invoke_map_dbl(
  .f,
  .x = list(NULL),
  ...,
  .env = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = deprecated()
)

future_invoke_map_int(
  .f,
  .x = list(NULL),
  ...,
  .env = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = deprecated()
)

future_invoke_map_lgl(
  .f,
  .x = list(NULL),
  ...,
  .env = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = deprecated()
)

future_invoke_map_dfr(
  .f,
  .x = list(NULL),
  ...,
  .env = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = deprecated()
)

future_invoke_map_dfc(
  .f,
  .x = list(NULL),
  ...,
  .env = NULL,
  .options = furrr_options(),
  .env_globals = parent.frame(),
  .progress = deprecated()
)

Arguments

.f
A list of functions.

.x
A list of argument-lists the same length as .f (or length 1). The default argument, list(NULL), will be recycled to the same length as .f, and will call each function with no arguments (apart from any supplied in ....)

...
Additional arguments passed to each function.

.env
Environment in which do.call() should evaluate a constructed expression. This only matters if you pass as .f the name of a function rather than its value, or as .x symbols of objects rather than their values.

.options
The future specific options to use with the workers. This must be the result from a call to furrr_options().

.env_globals
The environment to look for globals required by .f, ..., and .x.

.progress
: No longer has any effect. Use the progressr package instead.

Examples

# \donttest{
plan(multiprocess)
# }

df <- dplyr::tibble(
  f = c("runif", "rpois", "rnorm"),
  params = list(
    list(n = 10),
    list(n = 5, lambda = 10),
    list(n = 10, mean = -3, sd = 10)
  )
)

future_invoke_map(df$f, df$params)

#> [[1]]
#>  [1] 0.2526452 0.3531381 0.4066444 0.9433727 0.6455873 0.6931532 0.4960275
#>  [8] 0.6883092 0.9179617 0.1485430
#> 
#> [[2]]
#> [1] 10  7  3  7 12
#> 
#> [[3]]
#>  [1]   5.474396  -4.701210  -8.059296  -4.074736  20.395892  -1.649295
#>  [7] -10.359960  -3.345520  -3.414092  -7.097007
#> 

# \dontshow{
# Close open connections for R CMD Check
if (!inherits(plan(), "sequential")) plan(sequential)
# }