Apply a function to each element of a vector via futures

These functions work exactly the same as purrr::map() functions, but allow you to run the map in parallel. The documentation is adapted from both purrr::map(), and future.apply::future_lapply(), so look there for more details.

future_map(.x, .f, ..., .options = future_options(), .progress = deprecated())

future_map_chr(
  .x,
  .f,
  ...,
  .options = future_options(),
  .progress = deprecated()
)

future_map_dbl(
  .x,
  .f,
  ...,
  .options = future_options(),
  .progress = deprecated()
)

future_map_int(
  .x,
  .f,
  ...,
  .options = future_options(),
  .progress = deprecated()
)

future_map_lgl(
  .x,
  .f,
  ...,
  .options = future_options(),
  .progress = deprecated()
)

future_map_dfr(
  .x,
  .f,
  ...,
  .id = NULL,
  .options = future_options(),
  .progress = deprecated()
)

future_map_dfc(
  .x,
  .f,
  ...,
  .options = future_options(),
  .progress = deprecated()
)

future_map_if(
  .x,
  .p,
  .f,
  ...,
  .options = future_options(),
  .progress = deprecated()
)

future_map_at(
  .x,
  .at,
  .f,
  ...,
  .options = future_options(),
  .progress = deprecated()
)

future_walk(.x, .f, ..., .options = future_options())

Arguments

.x
A list or atomic vector.

.f
A function, formula, or vector (not necessarily atomic).

If a function, it is used as is.

If a formula, e.g. ~ .x + 2, it is converted to a function. There are three ways to refer to the arguments:

For a single argument function, use .

For a two argument function, use .x and .y

For more arguments, use ..1, ..2, ..3 etc

This syntax allows you to create very compact anonymous functions.

If character vector, numeric vector, or list, it is converted to an extractor function. Character vectors index by name and numeric vectors index by position; use a list to index by position and name at different levels. If a component is not present, the value of .default will be returned.

...
Additional arguments passed on to the mapped function.

.options
The future specific options to use with the workers. This must be the result from a call to future_options().

.progress
: No longer has any effect. Use the progressr package instead.

.id
Either a string or NULL. If a string, the output will contain a variable with that name, storing either the name (if .x is named) or the index (if .x is unnamed) of the input. If NULL, the default, no variable will be created.

Only applies to _dfr variant.

.p
A single predicate function, a formula describing such a predicate function, or a logical vector of the same length as .x. Alternatively, if the elements of .x are themselves lists of objects, a string indicating the name of a logical element in the inner lists. Only those elements where .p evaluates to TRUE will be modified.

.at
A character vector of names, positive numeric vector of positions to include, or a negative numeric vector of positions to exlude. Only those elements corresponding to .at will be modified. If the tidyselect package is installed, you can use vars() and the tidyselect helpers to select elements.

Value

All functions return a vector the same length as .x.

future_map() returns a list, future_map_lgl() a logical vector, future_map_int() an integer vector, future_map_dbl() a double vector, and future_map_chr() a character vector. The output of .f will be automatically typed upwards, e.g. logical -> integer -> double -> character.

Examples

library(furrr)
library(dplyr) # for the pipe

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# \donttest{
plan(multiprocess)
# }

1:10 %>%
  future_map(rnorm, n = 10) %>%
  future_map_dbl(mean)

#>  [1]  1.245161  2.119907  3.328149  3.629935  4.824969  6.583845  6.849823
#>  [8]  7.593139  8.481830 10.268107

# If each element of the output is a data frame, use
# future_map_dfr to row-bind them together:
mtcars %>%
  split(.$cyl) %>%
  future_map(~ lm(mpg ~ wt, data = .x)) %>%
  future_map_dfr(~ as.data.frame(t(as.matrix(coef(.)))))

#>   (Intercept)        wt
#> 1    39.57120 -5.647025
#> 2    28.40884 -2.780106
#> 3    23.86803 -2.192438

# You can be explicit about what gets exported to the workers

# To see this, use multisession (NOT multicore if on a Mac as the forked workers
# still have access to this environment)
# \donttest{
plan(multisession)
# }
x <- 1
y <- 2

# This will fail, y is not exported (no black magic occurs)
# future_map(1, ~y, .options = future_options(globals = "x"))

# y is exported
future_map(1, ~y, .options = future_options(globals = "y"))

#> [[1]]
#> [1] 2
#> 

# \dontshow{
# Close open connections for R CMD Check
if (!inherits(plan(), "sequential")) plan(sequential)
# }